[refactor] move all plan a trip utils functions in a utils file
